fix: move IL-1-importing pty_mode tests from core to execution test file
test_default_subprocess_runner_pty_mode_default_false and
test_run_managed_async_pty_mode_default_false imported from
autoskillit.execution.process (IL-1) but carried layer("core")
markers, making them invisible to CI when execution/process.py
changes. Moved both to tests/execution/test_process_submodules.py
where they inherit the correct layer("execution") marker.
test_subprocess_runner_protocol_pty_mode_default_false remains in
tests/core/test_types.py as it imports only from autoskillit.core.
