+++ tests/core/test_types.py
@@ -353,24 +353,6 @@ def test_subprocess_runner_protocol_pty_mode_default_false():
     assert sig.parameters["pty_mode"].default is False
 
 
-def test_default_subprocess_runner_pty_mode_default_false():
-    import inspect
-
-    from autoskillit.execution.process import DefaultSubprocessRunner
-
-    sig = inspect.signature(DefaultSubprocessRunner.__call__)
-    assert sig.parameters["pty_mode"].default is False
-
-
-def test_run_managed_async_pty_mode_default_false():
-    import inspect
-
-    from autoskillit.execution.process import run_managed_async
-
-    sig = inspect.signature(run_managed_async)
-    assert sig.parameters["pty_mode"].default is False
-
-
 # ---------------------------------------------------------------------------
 # CIRunScope event field
 # ---------------------------------------------------------------------------
