+++ tests/execution/test_process_submodules.py
@@ -103,3 +103,21 @@ def test_process_facade_reexports_all_public_symbols():
     assert len(process.__all__) >= 21, (
         f"process.py __all__ has {len(process.__all__)} symbols, expected at least 21"
     )
+
+
+def test_default_subprocess_runner_pty_mode_default_false():
+    import inspect
+
+    from autoskillit.execution.process import DefaultSubprocessRunner
+
+    sig = inspect.signature(DefaultSubprocessRunner.__call__)
+    assert sig.parameters["pty_mode"].default is False
+
+
+def test_run_managed_async_pty_mode_default_false():
+    import inspect
+
+    from autoskillit.execution.process import run_managed_async
+
+    sig = inspect.signature(run_managed_async)
+    assert sig.parameters["pty_mode"].default is False
